LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scrollbar-gutter-001-expected.txt

auto/auto
scroll/auto
visible/auto
hidden/auto
clip/auto
auto/stable
scroll/stable
visible/stable
hidden/stable
clip/stable
auto/stable both-edges
scroll/stable both-edges
visible/stable both-edges
hidden/stable both-edges
clip/stable both-edges

PASS overflow auto, scrollbar-gutter auto
PASS overflow scroll, scrollbar-gutter auto
PASS overflow visible, scrollbar-gutter auto
PASS overflow hidden, scrollbar-gutter auto
PASS overflow clip, scrollbar-gutter auto
PASS overflow auto, scrollbar-gutter stable
PASS overflow scroll, scrollbar-gutter stable
PASS overflow visible, scrollbar-gutter stable
PASS overflow hidden, scrollbar-gutter stable
PASS overflow clip, scrollbar-gutter stable
PASS overflow auto, scrollbar-gutter stable both-edges
PASS overflow scroll, scrollbar-gutter stable both-edges
PASS overflow visible, scrollbar-gutter stable both-edges
PASS overflow hidden, scrollbar-gutter stable both-edges
PASS overflow clip, scrollbar-gutter stable both-edges
